Refactor tests, reduce size_t usage within chain/message types.

+++ include/bitcoin/bitcoin/chain/transaction_input.hpp
@@ -67,9 +67,9 @@ public:
 
     bool is_final() const;
 
-    size_t satoshi_size() const;
+    uint64_t satoshi_size() const;
 
-    static size_t satoshi_fixed_size();
+    static uint64_t satoshi_fixed_size();
 
 private:
