+++ src/message/network_address.cpp
@@ -127,12 +127,12 @@ data_chunk network_address::to_data() const
     return result;
 }
 
-size_t network_address::satoshi_size() const
+uint64_t network_address::satoshi_size() const
 {
     return network_address::satoshi_fixed_size();
 }
 
-size_t network_address::satoshi_fixed_size()
+uint64_t network_address::satoshi_fixed_size()
 {
     return 26;
 }
